requirements.txt
cffi==1.11.5
hkdf==0.0.3
pycparser==2.19
pycrypto==2.6.1
PyNaCl==1.3.0
six==1.12.0
